requirements.txt
absl-py==2.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
filelock==3.13.1
Flask==3.1.0
fsspec==2024.2.0
gdown==5.2.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
keras==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.0
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.3
optree==0.13.1
packaging==24.2
pillow==10.2.0
Pygments==2.18.0
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
setuptools==75.1.0
soupsieve==2.6
sympy==1.13.1
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
typing_extensions==4.9.0
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.44.0
